friends_service/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.1.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-django==4.10.0
pytz==2025.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
